refactor: change name of the hook
