+++ packages/hooks/src/index.ts
@@ -1,6 +1,6 @@
-export { default as useAccountPlatformDetails } from './useAccountPlatformDetails';
 export { default as useAccountTransferVisible } from './useAccountTransferVisible';
 export { default as useCountdown } from './useCountdown';
+export { default as useCurrentAccountDetails } from './useCurrentAccountDetails';
 export { default as useDepositLocked } from './useDepositLocked';
 export { default as useHasActiveRealAccount } from './useHasActiveRealAccount';
 export { default as useHasMaltaInvestAccount } from './useHasMaltaInvestAccount';
