+++ packages/hooks/src/useCurrentAccountDetails.ts
@@ -1,10 +1,10 @@
 import { useStore } from '@deriv/stores';
 
-const useAccountPlatformDetails = () => {
+const useCurrentAccountDetails = () => {
     const { client } = useStore();
     const { account_list, loginid } = client;
 
     return account_list.find(account => loginid === account.loginid);
 };
 
-export default useAccountPlatformDetails;
+export default useCurrentAccountDetails;
